% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcesteqn.R
\name{mcesteqn}
\alias{mcesteqn}
\title{Generate and Correct Generalized Estimating Equations}
\usage{
mcesteqn(beta, lb, n, m, X, mcov, ind, Y, mult)
}
\arguments{
\item{beta}{The value of covariates}

\item{lb}{the dimension of covariates}

\item{n}{the total number of observations}

\item{m}{the number of observations for each subject}

\item{X}{model matrix for the covariates for each ID}

\item{mcov}{the covariance matrix for the surrogate variables}

\item{ind}{the index of the surrogate covariates}

\item{Y}{the response variable vector for each ID}

\item{mult}{mutliplier}
}
\description{
The \code{mcesteqn} function generates the corrected generalized
estimating equations for each subject and each time point. Then the corrected
estimating equations are combined using General Method of Moments (GMM) for
each time point.
}
\details{
This function is used in \code{mcgmm} function to get the estimates.
}
