% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcesteqn.R
\name{mcgmm.control}
\alias{mcgmm.control}
\title{Estimation via Generalized Method of Moments Approach with
Measurement Errors Corrected}
\usage{
mcgmm.control(epsilon = 1e-08, trace = FALSE, maxit = 150)
}
\arguments{
\item{epsilon}{The relative steplength tolerance; the default value is 1e-8}

\item{trace}{A logical variable indicating if detailed report of the progress
of iternation is given}

\item{maxit}{The maximum iternation number of major iternations.
See corresponding documentation to \code{\link{nleqslv}{nleqslv}}}
}
\description{
Auxiliary function for \code{mcgmm} function.
}
\details{
Only used internally in \code{mcgmm}
}
