% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcesteqn.R
\name{mcgmm}
\alias{mcgmm}
\title{Estimation via Generalized Method of Moments Approach with
Measurement Errors Corrected}
\usage{
mcgmm(
  formula,
  data,
  me.var,
  mcov,
  time.var,
  id.var,
  init.beta,
  family = "binomial",
  control = list(),
  modify_inv = FALSE,
  finsam_cor = TRUE
)
}
\arguments{
\item{formula}{a symbolic description of the model}

\item{data}{a data frame that contains variables in the model
and corresponding a variables identifying subjects and time points}

\item{me.var}{names of variables with measurement errors in the data}

\item{mcov}{list of the covariance matrices for the surrogate variables}

\item{time.var}{name of variable that identifies different time points in
the data}

\item{id.var}{name of variable that identifies clusters in the data}

\item{init.beta}{The initial guess of the estimates}

\item{family}{the family of response variable}

\item{control}{a list of parameters that pass into the estimating process}

\item{modify_inv}{a logical variable specifying whether a not invertible
matrix should be fixed}

\item{finsam_cor}{a logical variable specifying whether or not the finite
sample bias should be corrected}
}
\description{
The \code{mcgmm} function returns the estimation by combining
and solving measurement errors corrected estimating equations via generalized
method of moments.
}
\details{
The input of data must be a data.frame.
}
