% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcesteqn.R
\name{process_eivgmm}
\alias{process_eivgmm}
\title{Data processing for \code{eivgmm} and \code{eivgmm_R} function}
\usage{
process_eivgmm(formula, data, me.var, time.var, id.var)
}
\arguments{
\item{formula}{a symbolic description of the model}

\item{data}{a data frame that contains variables in the model
and corresponding a variables identifying subjects and time points}

\item{me.var}{names of variables with measurement errors in the data}

\item{time.var}{name of variable that identifies different time points in
the data}

\item{id.var}{name of variable that identifies clusters in the data}
}
\description{
This function is built to process data in a way that all
the obversations are grouped by the time variable.
}
