% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcesteqn.R
\name{inference_rcpp}
\alias{inference_rcpp}
\title{Produce Inference Terms}
\usage{
inference_rcpp(
  lb,
  m,
  n,
  X,
  Y,
  beta,
  mcov,
  ind,
  maxit,
  eig_tol,
  conv_tol,
  modify_inv
)
}
\arguments{
\item{lb}{the dimension of covariates}

\item{m}{the number of observations for each subject}

\item{n}{the total number of observations}

\item{X}{model matrix for the covariates for each ID}

\item{Y}{the response variable vector for each ID}

\item{beta}{the stimated covariates}

\item{mcov}{the covariance matrix for the surrogate variables}

\item{ind}{the index of the surrogate covariates}

\item{maxit}{maxit variable for nearPD function}

\item{eig_tol}{eigenvalue tolerance variable for nearPD function}

\item{conv_tol}{convergence tolerance for nearPD function}

\item{modify_inv}{a logical variable specifying whether a not invertible
matrix should be fixed}
}
\description{
Produces inference terms like asymptotic variance.
}
