% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcesteqn.R
\name{mcesteqn_rcpp}
\alias{mcesteqn_rcpp}
\title{Generate and Correct Generalized Estimating Equations via Rcpp}
\usage{
mcesteqn_rcpp(lb, m, n, X, Y, beta, mcov, ind, modify_vinv)
}
\arguments{
\item{lb}{the dimension of covariates}

\item{m}{the number of observations for each subject}

\item{n}{the total number of observations}

\item{X}{model matrix for the covariates for each ID}

\item{Y}{the response variable vector for each ID}

\item{beta}{The value of covariates}

\item{mcov}{the covariance matrix for the surrogate variables}

\item{ind}{the index of the surrogate covariates}

\item{modify_vinv}{a logical variable indicating whether or not the matrix
should be modified for the inversion}
}
\description{
The \code{mcesteqn_rcpp} function is the rcpp version of the
function \code{mcesteqn}.
}
